Move cli/util update checks to cli/cloud package.
